Created custom structs for rects and vectors
Introduced custom structs to underlie the screen space math. Everything is now based on Integers that represents 10000s of screen space. (0,0) would be the upper left corner while (10000,10000) would be the lower right. The use of custom structs without extra properties hopefully avoids the recursive issue we saw with the built in Vector classes.