Updated MetaDataTracker
Added a few new features to the MetaDataTracker.

Both KeyFrameRate and TweenFrameRate can be set independently rather than being locked to an update loop. I also added a setting to pick which Update function to use for tracking. Ultimately we may not keep this as a feature but we can test with it.

Tween Frames will now only be created if there are Tween objects tracked that have information to report. If no Tween objects come back in the call then a frame won't be snapped. We can see how this impacts bandwidth and accounting on the remote end.

Some various other reorganizing and renaming of properties.