First Person Toy Adjustments
Changed how screenRects are calculated to use WorldToViewportPoint instead of WorldToScreenPoint, which provides depth information as well as position. ScreenRects are now DepthRects, which have a z value that denotes distance from camera, with a negative value meaning it is behind the view.

Added a system for both the MetaDataTracker and individual IMetaDataTrackables to reference different camera to calculate screenRects, when not set it defaults to Camera.main.

Added a BasicTrackable that can be easily added to any object to just report a screenRect for testing.